+++ src/features/help.ts
@@ -7,9 +7,10 @@ export class Help {
   /**
    * Shows help message
    * @param ctx Text message context
+   * @param cleanCommand Clean command name
    */
-  public async command(ctx: TextMessageCtx): Promise<void> {
-    if (!isCleanCommand("help", ctx.message.text) && !isCleanCommand("start", ctx.message.text)) {
+  public async command(ctx: TextMessageCtx, cleanCommand: string): Promise<void> {
+    if (!isCleanCommand(cleanCommand, ctx.message.text)) {
       return; // Not clean command, ignore.
     }