+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.9.x (2023-x-x)
 
+- Improvement: descriptions of the features
 - Improvement: pagination on time zone settings
 - Fix: profanity filter ignored message which starts from command
 - Fix: profanity filter ignored profanity in forwarded names